Sema: Factor out SolverTrail from ConstraintGraph
